dakboard/requirements.txt
pytest
pillow
python-dotenv
requests